requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
groq>=0.9.0
